move wallet.bin to vm/templates and track with git-lfs
