+++ api/grpcserver/grpcserver_test.go
@@ -52,7 +52,6 @@ import (
 	"github.com/spacemeshos/go-spacemesh/system"
 	"github.com/spacemeshos/go-spacemesh/txs"
 	"github.com/spacemeshos/go-spacemesh/vm"
-	walletProgram "github.com/spacemeshos/go-spacemesh/vm/programs/wallet"
 	"github.com/spacemeshos/go-spacemesh/vm/sdk"
 	"github.com/spacemeshos/go-spacemesh/vm/sdk/wallet"
 	walletTemplate "github.com/spacemeshos/go-spacemesh/vm/templates/wallet"
@@ -2397,7 +2396,7 @@ func TestVMAccountUpdates(t *testing.T) {
 	// add the wallet template account
 	accounts[len(accounts)-1] = types.Account{
 		Address:         walletTemplate.TemplateAddress,
-		State:           walletProgram.PROGRAM,
+		State:           walletTemplate.PROGRAM,
 		TemplateAddress: &walletTemplate.TemplateAddress,
 	}
 	require.NoError(t, svm.ApplyGenesis(accounts))
