+++ config/presets/fastnet.go
@@ -10,8 +10,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/activation"
 	"github.com/spacemeshos/go-spacemesh/common/types"
 	"github.com/spacemeshos/go-spacemesh/config"
-	"github.com/spacemeshos/go-spacemesh/vm/programs/wallet"
-	walletTemplate "github.com/spacemeshos/go-spacemesh/vm/templates/wallet"
+	"github.com/spacemeshos/go-spacemesh/vm/templates/wallet"
 )
 
 func init() {
@@ -47,10 +46,10 @@ func fastnet() config.Config {
 	conf.Genesis = config.GenesisConfig{
 		ExtraData: "fastnet",
 		Accounts: map[string]uint64{
-			walletTemplate.TemplateAddress.StringWithHRP(conf.NetworkHRP): 10000000000000,
+			wallet.TemplateAddress.StringWithHRP(conf.NetworkHRP): 10000000000000,
 		},
 		Templates: map[string][]byte{
-			walletTemplate.TemplateAddress.StringWithHRP(conf.NetworkHRP): wallet.PROGRAM,
+			wallet.TemplateAddress.StringWithHRP(conf.NetworkHRP): wallet.PROGRAM,
 		},
 	}
