+++ vm/templates/wallet/program.go
