+++ vm/templates/wallet/wallet.bin
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8daf2bd314b9bb263139c4e9d8e9e22f28ddfef8771cd1b4e2792ff32a6254c4
+size 147924
